+++ src/main/java/com/stock/flex/config/JwtAuthFilter.java
@@ -3,7 +3,7 @@ package com.stock.flex.config;
 import java.io.IOException;
 import java.util.Date;
 
-import com.stock.flex.security.service.JwtService;
+import com.stock.flex.security.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
