+++ src/main/java/com/stock/flex/resource/AuthResource.java
@@ -3,7 +3,7 @@ package com.stock.flex.resource;
 import com.stock.flex.resource.request.AuthRequest;
 import com.stock.flex.resource.response.AuthResponse;
 import com.stock.flex.resource.request.RegisterRequest;
-import com.stock.flex.security.service.AuthService;
+import com.stock.flex.security.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
