+++ src/main/java/com/stock/flex/resource/StockResource.java
@@ -106,5 +106,4 @@ StockRepository repository;
         return ResponseEntity.ok().build();
     }
 
-
 }
