+++ src/main/java/com/stock/flex/resource/UserResource.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 
 import com.stock.flex.resource.response.UserResponse;
 import com.stock.flex.entity.UserEntity;
-import com.stock.flex.security.service.UserService;
+import com.stock.flex.security.UserService;
 import io.swagger.v3.oas.annotations.Hidden;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -26,20 +26,20 @@ public class UserResource {
 	@GetMapping
 	public ResponseEntity<List<UserResponse>> findAll() {
 		final List<UserEntity> personEntities = service.findAll();
-		final List<UserResponse> dtos = personEntities.stream().map(p -> new UserResponse(p)).toList();
-		return ResponseEntity.ok(dtos);
+		final List<UserResponse> responses = personEntities.stream().map(UserResponse::new).toList();
+		return ResponseEntity.ok(responses);
 	}
 
 	@PostMapping
-	public ResponseEntity<UserResponse> create(@RequestBody UserResponse dto) {
-		dto.setPassword(passwordEncoder.encode(dto.getPassword()));
-		return ResponseEntity.ok(service.create(dto));
+	public ResponseEntity<UserResponse> create(@RequestBody UserResponse response) {
+		response.setPassword(passwordEncoder.encode(response.getPassword()));
+		return ResponseEntity.ok(service.create(response));
 	}
 
 	@PutMapping
-	public ResponseEntity<UserResponse> update(@RequestBody UserResponse dto) {
-		dto.setPassword(passwordEncoder.encode(dto.getPassword()));
-		return ResponseEntity.ok(service.create(dto));
+	public ResponseEntity<UserResponse> update(@RequestBody UserResponse response) {
+		response.setPassword(passwordEncoder.encode(response.getPassword()));
+		return ResponseEntity.ok(service.create(response));
 	}
 	
 	@DeleteMapping("/{id}")
