+++ src/main/resources/db/migration/V2__create-table-stock.sql
@@ -1,7 +1,8 @@
-ALTER TABLE product ADD (
-    expiration_date DATE NOT NULL,
-    purchase_date DATE NOT NULL,
-    brand VARCHAR(255) NOT NULL,
-    purchase_location VARCHAR(255) NOT NULL,
-    has_coupon BOOLEAN NOT NULL
+ALTER TABLE product
+ADD (
+    expirationDate DATE,
+    purchaseDate DATE,
+    brand VARCHAR(255),
+    purchaseLocation VARCHAR(255),
+    hasCoupon BOOLEAN
 );
